refacto(crate): split into lib + bin for downstream bundling
